indirizzo

+++ src/main/java/it/epicode/bwfinalboss/repository/IndirizzoRepository.java
@@ -6,5 +6,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IndirizzoRepository extends JpaRepository<Indirizzo, Integer> {
+public interface IndirizzoRepository extends JpaRepository<Indirizzo, Long> {
 }
